old_code_soon_deleted/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=2.1.0=pypi_0
accelerate=0.25.0=pypi_0
aiohttp=3.8.6=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.2.0=pypi_0
appdirs=1.4.4=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
babel=2.14.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
beautifulsoup4=4.12.3=pypi_0
bert-score=0.3.13=pypi_0
bertopic=0.16.0=pypi_0
biopython=1.81=pypi_0
bitsandbytes=0.42.0=pypi_0
blas=1.0=openblas
bleach=6.1.0=pypi_0
blis=0.7.11=pypi_0
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.7.22=hf0a4a13_0
catalogue=2.0.10=pypi_0
certifi=2023.7.22=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.1=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.16.0=pypi_0
comm=0.1.4=pyhd8ed1ab_0
confection=0.1.3=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
cython=0.29.37=pypi_0
dataclasses-json=0.6.3=pypi_0
datasets=2.14.6=pypi_0
debugpy=1.6.7=py310h313beb8_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
dill=0.3.7=pypi_0
distro=1.9.0=pypi_0
dnspython=2.5.0=pypi_0
en-core-web-sm=3.7.0=pypi_0
evaluate=0.4.1=pypi_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.5.0=h313beb8_0
fastjsonschema=2.19.1=pypi_0
fasttext=0.9.2=pypi_0
filelock=3.13.0=pypi_0
fonttools=4.43.1=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.10.0=pypi_0
fst-pso=1.8.1=pypi_0
fuzzytm=2.0.5=pypi_0
gensim=4.3.2=pypi_0
h11=0.14.0=pypi_0
hdbscan=0.8.33=pypi_0
httpcore=1.0.2=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.20.1=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
ipykernel=6.26.0=pyh3cd1d5f_0
ipython=8.16.1=pyh31c8845_0
ipywidgets=8.1.2=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
json5=0.9.14=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-events=0.9.0=pypi_0
jupyter-lsp=2.2.2=pypi_0
jupyter-server=2.12.5=pypi_0
jupyter-server-terminals=0.5.2=pypi_0
jupyter_client=8.5.0=pyhd8ed1ab_0
jupyter_core=5.3.0=py310hca03da5_0
jupyterlab=4.1.1=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.25.3=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
kiwisolver=1.4.5=pypi_0
langchain=0.1.0=pypi_0
langchain-community=0.0.11=pypi_0
langchain-core=0.1.10=pypi_0
langchain-openai=0.0.2=pypi_0
langcodes=3.3.0=pypi_0
langsmith=0.0.80=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.21=h269037a_0
libsodium=1.0.18=h27ca646_1
llvm-openmp=14.0.6=hc6e5704_0
llvmlite=0.42.0=pypi_0
loguru=0.7.2=pypi_0
markupsafe=2.1.3=pypi_0
marshmallow=3.20.2=pypi_0
matplotlib=3.8.0=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
miniful=0.0.6=pypi_0
mistune=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
murmurhash=1.0.10=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclient=0.9.0=pypi_0
nbconvert=7.16.0=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
nltk=3.8.1=pypi_0
notebook=7.1.0=pypi_0
notebook-shim=0.2.4=pypi_0
numba=0.59.0=pypi_0
numpy=1.26.1=pypi_0
numpy-base=1.26.0=py310ha9811e2_0
openai=1.7.1=pypi_0
openssl=3.1.4=h0d3ecfb_0
overrides=7.7.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.4=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.1.0=pypi_0
pinecone-client=2.2.4=pypi_0
pip=23.3=py310hca03da5_0
platformdirs=3.11.0=pyhd8ed1ab_0
plotly=5.19.0=pypi_0
portalocker=2.8.2=pypi_0
preshed=3.0.9=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.0=py310h1a28f6b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=13.0.0=pypi_0
pybind11=2.11.1=pypi_0
pycparser=2.21=pypi_0
pydantic=2.4.2=pypi_0
pydantic-core=2.10.1=pypi_0
pyfume=0.2.25=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pynndescent=0.5.11=pypi_0
pyparsing=3.1.1=pypi_0
pyphen=0.14.0=pypi_0
pysbd=0.3.4=pypi_0
python=3.10.13=hb885b13_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.1=pypi_0
python-json-logger=2.0.7=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.0=py310h313beb8_0
ragas=0.1.2=pypi_0
rank-bm25=0.2.2=pypi_0
readline=8.2=h1a28f6b_0
referencing=0.33.0=pypi_0
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
responses=0.18.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rouge-score=0.1.2=pypi_0
rpds-py=0.18.0=pypi_0
safetensors=0.4.1=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.3=py310h20cbe94_0
seaborn=0.13.0=pypi_0
send2trash=1.8.2=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.0.0=py310hca03da5_0
simpful=2.11.0=pypi_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0.post10=pypi_0
smart-open=6.4.0=pypi_0
smart_open=5.2.1=py310hca03da5_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
spacy=3.7.2=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlalchemy=2.0.25=pypi_0
sqlite=3.41.2=h80987f9_0
srsly=2.4.8=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.2.3=pypi_0
terminado=0.18.0=pypi_0
textstat=0.7.3=pypi_0
thinc=8.2.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tiktoken=0.5.2=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.15.0=pypi_0
tomli=2.0.1=pypi_0
torch=2.1.0=pypi_0
torchdata=0.7.1=pypi_0
torchtext=0.16.1=pypi_0
torchvision=0.16.0=pypi_0
tornado=6.3.3=py310h80987f9_0
tqdm=4.66.1=pypi_0
traitlets=5.12.0=pyhd8ed1ab_0
transformers=4.36.2=pypi_0
typer=0.9.0=pypi_0
types-python-dateutil=2.8.19.20240106=pypi_0
typing-extensions=4.8.0=hd8ed1ab_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023.3=pypi_0
umap-learn=0.5.5=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.0.7=pypi_0
wasabi=1.1.2=pypi_0
wcwidth=0.2.8=pyhd8ed1ab_0
weasel=0.3.3=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.7.0=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.10=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.2=h80987f9_0
yarl=1.9.2=pypi_0
zeromq=4.3.4=hc377ac9_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5a0b063_0
